addon/lvgl: Use CDisplay instead of CBcmFrameBuffer
By handing over a pointer to a CDisplay instance to the constructor
we can show the GUI on any display, which supports this generic
interface. We generally use the LVGL mouse cursor now to be able
to support the cursor on displays, for which the firmware does not
provide mouse cursor support. The constructor does not need a pointer
to CInterruptSystem any more, because the display flush is handled
in the display driver now.