net/tftpdaemon: Add IsAccessAllowed()
Allows to check the permission based on the IP address of the client,
the requested filename and the request type. By default any access is
allowed. Can be overridden in the derived class to implement access
control. Also make the member variables protected for implementing
special functions in the derived class.

Issue #514