input/mouse: Support mouse in a window display everywhere
Before the mouse window display had to have the origin 0,0 on parent
display. Now it can be everywhere. This requires a modification to
CMouseDevice::Setup(), which gets a pointer to the display now,
instead of display width and height. sample/10-usbmouse, addon/lvgl
and addon/ugui have been modified to use the changed API. CDisplay
can be requested for its parent display and origin now.