sound/pcm512xsoundcontroller: Improve detection of PCM512x
Before it was only tested, if an I2C device is at address 0x4C, 0x4D or
at the specified address (without auto-probing). This did not work, when
the address 0x1A was specified for the WM8960, which also responded at
the given address. Because the PCM512x was probed first, the driver
thought, it was there and initialized for PCM512x use, even when a
WM8960 was connected. Now a register is read via I2C and checked for its
reset value. Only when it matches, the PCM512x is detected.

See: #518